backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
pymysql==1.1.0
python-dotenv==1.0.1
cryptography==44.0.2
pydantic[email]>=2.5
pytest==8.1.1